requirements.txt
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==22.2.0
bcrypt==4.0.1
bson==0.5.10
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.89.1
greenlet==2.0.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
psycopg2==2.9.5
py==1.11.0
pyasn1==0.4.8
pydantic==1.10.4
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
redis==4.5.1
requests==2.26.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.0
starlette==0.22.0
toml==0.10.2
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
numpy
clickhouse-connect
pymongo==4.4.0
pyOpenSSL==23.2.0
proxmoxer==2.0.1
paramiko==3.2.0
python_wireguard==0.2.2